cleaner README, refactored to put the new layer in its own module
